+++ metricflow/dataflow/dataflow_plan.py
@@ -34,7 +34,7 @@ from metricflow.dataflow.builder.partitions import (
 from metricflow.dataflow.sql_table import SqlTable
 from metricflow.dataset.dataset import DataSet
 from metricflow.model.objects.metric import CumulativeMetricWindow
-from metricflow.model.objects.elements.measure import AggregationType
+from metricflow.aggregation_properties import AggregationType
 from metricflow.object_utils import pformat_big_objects
 from metricflow.references import TimeDimensionReference
 from metricflow.specs import (
