+++ metricflow/test/model/parsing/test_data_source_parsing.py
@@ -1,9 +1,9 @@
 import textwrap
 
+from metricflow.aggregation_properties import AggregationType
 from metricflow.model.objects.common import YamlConfigFile
 from metricflow.model.objects.data_source import DataSourceOrigin, MutabilityType
 from metricflow.model.objects.elements.identifier import IdentifierType
-from metricflow.model.objects.elements.measure import AggregationType
 from metricflow.model.objects.elements.dimension import DimensionType
 from metricflow.model.parsing.dir_to_model import parse_yaml_files_to_model
 from metricflow.time.time_granularity import TimeGranularity
